Example 5.6 and 5.7 added
Example 5.6 and 5.7 is added, including figure 5.16 and 5.17.

Note: Since the noise is combined of noises of different periods, this
method does not do as expected on this SPECIFIC image. Nevertheless, it
is reasonable.

TODO: replace this image by some other appropriate image, the result
shall be the same as Gonzalez.